remove ContentNode subclasses (Entity + Repository + DataPersister)
