--- api/src/DataPersister/ContentNode/ContentNodeAbstractDataPersister.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace App\DataPersister\ContentNode;
-
-use App\DataPersister\Util\AbstractDataPersister;
-use App\Entity\ContentNode;
-
-abstract class ContentNodeAbstractDataPersister extends AbstractDataPersister {
-    /**
-     * @template T of ContentNode
-     *
-     * @param T $data
-     *
-     * @return T
-     */
-    public function beforeCreate($data): ContentNode {
-        // set root from parent
-        $data->parent->addChild($data);
-        $data->parent->root->addRootDescendant($data);
-
-        return $data;
-    }
-}
