--- api/src/DataPersister/ContentNode/MaterialNodeDataPersister.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\DataPersister\ContentNode;
-
-use App\DataPersister\Util\DataPersisterObservable;
-use App\Entity\ContentNode\MaterialNode;
-
-class MaterialNodeDataPersister extends ContentNodeAbstractDataPersister {
-    /**
-     * @throws \ReflectionException
-     */
-    public function __construct(
-        DataPersisterObservable $dataPersisterObservable
-    ) {
-        parent::__construct(
-            MaterialNode::class,
-            $dataPersisterObservable
-        );
-    }
-}
