--- api/src/DataPersister/ContentNode/SingleTextDataPersister.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\DataPersister\ContentNode;
-
-use App\DataPersister\Util\DataPersisterObservable;
-use App\Entity\ContentNode\SingleText;
-
-class SingleTextDataPersister extends ContentNodeAbstractDataPersister {
-    /**
-     * @throws \ReflectionException
-     */
-    public function __construct(
-        DataPersisterObservable $dataPersisterObservable
-    ) {
-        parent::__construct(
-            SingleText::class,
-            $dataPersisterObservable
-        );
-    }
-}
