--- api/src/DataPersister/ContentNode/StoryboardDataPersister.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\DataPersister\ContentNode;
-
-use App\DataPersister\Util\DataPersisterObservable;
-use App\Entity\ContentNode\Storyboard;
-
-class StoryboardDataPersister extends ContentNodeAbstractDataPersister {
-    /**
-     * @throws \ReflectionException
-     */
-    public function __construct(
-        DataPersisterObservable $dataPersisterObservable
-    ) {
-        parent::__construct(
-            Storyboard::class,
-            $dataPersisterObservable
-        );
-    }
-}
