+++ api/src/DataPersister/ContentNodeDataPersister.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace App\DataPersister;
+
+use App\DataPersister\Util\AbstractDataPersister;
+use App\DataPersister\Util\DataPersisterObservable;
+use App\Entity\ContentNode;
+
+class ContentNodeDataPersister extends AbstractDataPersister {
+    /**
+     * @throws \ReflectionException
+     */
+    public function __construct(
+        DataPersisterObservable $dataPersisterObservable
+    ) {
+        parent::__construct(
+            ContentNode::class,
+            $dataPersisterObservable
+        );
+    }
+
+    /**
+     * @param ContentNode $data
+     */
+    public function beforeCreate($data): ContentNode {
+        // set root from parent
+        $data->parent->addChild($data);
+        $data->parent->root->addRootDescendant($data);
+
+        return $data;
+    }
+}
