+++ api/src/Entity/BelongsToContentNodeTreeInterface.php
@@ -2,8 +2,6 @@
 
 namespace App\Entity;
 
-use App\Entity\ContentNode\ColumnLayout;
-
 interface BelongsToContentNodeTreeInterface {
-    public function getRoot(): ?ColumnLayout;
+    public function getRoot(): ?ContentNode;
 }
