--- api/src/Entity/ContentNode/ColumnLayout.php
@@ -1,126 +1,0 @@
-<?php
-
-namespace App\Entity\ContentNode;
-
-use ApiPlatform\Core\Annotation\ApiProperty;
-use ApiPlatform\Core\Annotation\ApiResource;
-use App\Entity\ContentNode;
-use App\Repository\ColumnLayoutRepository;
-use App\Util\EntityMap;
-use App\Validator\AssertJsonSchema;
-use App\Validator\ColumnLayout\AssertColumWidthsSumTo12;
-use App\Validator\ColumnLayout\AssertNoOrphanChildren;
-use App\Validator\ColumnLayout\ColumnLayoutPatchGroupSequence;
-use App\Validator\ColumnLayout\ColumnLayoutPostGroupSequence;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Serializer\Annotation\Groups;
-
-#[ApiResource(
-    routePrefix: '/content_node',
-    collectionOperations: [
-        'get' => [
-            'security' => 'is_authenticated()',
-        ],
-        'post' => [
-            'denormalization_context' => ['groups' => ['write', 'create']],
-            'security_post_denormalize' => 'is_granted("CAMP_MEMBER", object) or is_granted("CAMP_MANAGER", object)',
-            'validation_groups' => ColumnLayoutPostGroupSequence::class,
-        ],
-    ],
-    itemOperations: [
-        'get' => ['security' => 'is_granted("CAMP_COLLABORATOR", object) or is_granted("CAMP_IS_PROTOTYPE", object)'],
-        'patch' => [
-            'denormalization_context' => ['groups' => ['write', 'update']],
-            'security' => 'is_granted("CAMP_MEMBER", object) or is_granted("CAMP_MANAGER", object)',
-            'validation_groups' => ColumnLayoutPatchGroupSequence::class,
-        ],
-        'delete' => ['security' => '(is_granted("CAMP_MEMBER", object) or is_granted("CAMP_MANAGER", object)) and object.parent !== null'], // disallow delete when contentNode is a root node
-    ],
-    denormalizationContext: ['groups' => ['write']],
-    normalizationContext: ['groups' => ['read']],
-)]
-#[ORM\Entity(repositoryClass: ColumnLayoutRepository::class)]
-#[ORM\Table(name: 'content_node_columnlayout')]
-class ColumnLayout extends ContentNode {
-    public const COLUMNS_SCHEMA = [
-        'type' => 'array',
-        'items' => [
-            'type' => 'object',
-            'additionalProperties' => false,
-            'required' => ['slot', 'width'],
-            'properties' => [
-                'slot' => [
-                    'type' => 'string',
-                    'pattern' => '^[1-9][0-9]*$',
-                ],
-                'width' => [
-                    'type' => 'integer',
-                    'minimum' => 3,
-                    'maximum' => 12,
-                ],
-            ],
-        ],
-    ];
-
-    /**
-     * JSON configuration for columns.
-     */
-    #[ApiProperty(example: [['slot' => '1', 'width' => 12]])]
-    #[Groups(['read', 'write'])]
-    #[AssertJsonSchema(schema: ColumnLayout::COLUMNS_SCHEMA, groups: ['columns_schema'])]
-    #[AssertColumWidthsSumTo12]
-    #[AssertNoOrphanChildren]
-    #[ORM\Column(type: 'json', nullable: true)]
-    public ?array $columns = [['slot' => '1', 'width' => 12]];
-
-    /**
-     * All content nodes that are part of this content node tree.
-     */
-    #[ApiProperty(readable: false, writable: false)]
-    #[ORM\OneToMany(targetEntity: ContentNode::class, mappedBy: 'root')]
-    public Collection $rootDescendants;
-
-    public function __construct() {
-        parent::__construct();
-        $this->rootDescendants = new ArrayCollection();
-    }
-
-    /**
-     * @return ContentNode[]
-     */
-    public function getRootDescendants(): array {
-        return $this->rootDescendants->getValues();
-    }
-
-    public function addRootDescendant(ContentNode $rootDescendant): self {
-        if (!$this->rootDescendants->contains($rootDescendant)) {
-            $this->rootDescendants[] = $rootDescendant;
-            $rootDescendant->root = $this;
-        }
-
-        return $this;
-    }
-
-    public function removeRootDescendant(ContentNode $rootDescendant): self {
-        if ($this->rootDescendants->removeElement($rootDescendant)) {
-            // reset the owning side (unless already changed)
-            if ($rootDescendant->root === $this) {
-                $rootDescendant->root = null;
-            }
-        }
-
-        return $this;
-    }
-
-    /**
-     * @param ColumnLayout $prototype
-     * @param EntityMap    $entityMap
-     */
-    public function copyFromPrototype($prototype, $entityMap): void {
-        parent::copyFromPrototype($prototype, $entityMap);
-
-        $this->columns = $prototype->columns;
-    }
-}
