--- api/src/Entity/ContentNode/SingleText.php
@@ -1,54 +1,0 @@
-<?php
-
-namespace App\Entity\ContentNode;
-
-use ApiPlatform\Core\Annotation\ApiResource;
-use App\Entity\ContentNode;
-use App\InputFilter;
-use App\Repository\SingleTextRepository;
-use App\Util\EntityMap;
-use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Serializer\Annotation\Groups;
-
-#[ApiResource(
-    routePrefix: '/content_node',
-    collectionOperations: [
-        'get' => [
-            'security' => 'is_authenticated()',
-        ],
-        'post' => [
-            'denormalization_context' => ['groups' => ['write', 'create']],
-            'security_post_denormalize' => 'is_granted("CAMP_MEMBER", object) or is_granted("CAMP_MANAGER", object)',
-            'validation_groups' => ['Default', 'create'],
-        ],
-    ],
-    itemOperations: [
-        'get' => ['security' => 'is_granted("CAMP_COLLABORATOR", object) or is_granted("CAMP_IS_PROTOTYPE", object)'],
-        'patch' => [
-            'denormalization_context' => ['groups' => ['write', 'update']],
-            'security' => 'is_granted("CAMP_MEMBER", object) or is_granted("CAMP_MANAGER", object)',
-            'validation_groups' => ['Default', 'update'],
-        ],
-        'delete' => ['security' => '(is_granted("CAMP_MEMBER", object) or is_granted("CAMP_MANAGER", object)) and object.parent !== null'], // disallow delete when contentNode is a root node
-    ],
-    denormalizationContext: ['groups' => ['write']],
-    normalizationContext: ['groups' => ['read']],
-)]
-#[ORM\Entity(repositoryClass: SingleTextRepository::class)]
-#[ORM\Table(name: 'content_node_singletext')]
-class SingleText extends ContentNode {
-    #[InputFilter\CleanHTML]
-    #[Groups(['read', 'write'])]
-    #[ORM\Column(type: 'text', nullable: true)]
-    public ?string $text = null;
-
-    /**
-     * @param SingleText $prototype
-     * @param EntityMap  $entityMap
-     */
-    public function copyFromPrototype($prototype, $entityMap): void {
-        parent::copyFromPrototype($prototype, $entityMap);
-
-        $this->text = $prototype->text;
-    }
-}
